refactor: seperate image components as new file
